*: move JoinProcessErrors and add test case
